[CHORE] create comments.py in app.views
